Avoid double builds for PR submissions
This avoids a build for the branch and a build for the PR, while
still allowing tagged builds for releases.

In general, we do not push directly to branches in the kopia/kopia
repo. However, besides tagged releases, there may be rare exceptions
such as minor edits to documentation via the Github Web UI.