fix(server): avoid repodiag blob upload failures
The kopia server never uploads logs to the repository,
because the Context used for uploading the log blobs
is already canceled when the PUT operation is initiated.

The context used for blob uploads is passed to
`repodiag.NewLogManager` when opening the repository.

In the case of the kopia server, the repository is asynchronously
opened in `server.Server.InitReposotoryAsync`. The context
passed to `repo.Open` is canceled after the "open repository"
server task completes. This issue was introduced in #1691

Change:
Use `context.WithoutCancel()` instead of the context passed
when the repo diagnoser is created.